Bug fix and simplification
